requirements.txt
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.0
aiohttp==3.9.5
colorama==0.4.6
httpx==0.25.2
plotly
streamlit
gdown
scipy
httpx